refactor(auth): Reorganize routers and services into VSA features
