+++ auth/src/features/api/mod.rs
@@ -1,0 +1,10 @@
+use axum::Router;
+
+use crate::{internal::config::Config, state::AppState};
+
+pub mod v1;
+
+pub fn router(config: &Config) -> Router<AppState> {
+    let route = Router::new().merge(v1::router(config));
+    Router::new().nest("/api", route)
+}
