+++ auth/src/features/api/v1/auth/verify_email.rs
@@ -1,3 +1,4 @@
+use axum::extract::State;
 use redis::aio::MultiplexedConnection;
 use sea_orm::DatabaseConnection;
 use serde::Deserialize;
@@ -6,7 +7,37 @@ use token::{TokenStore, services::RegisterTokenService};
 use crate::{
     internal::error::{AppError, AppErrorKind},
     models::users,
+    routers::{
+        extractor::Json,
+        response::app_error_to_response,
+        serializer::{Response, ResponseCode},
+    },
+    state::AppState,
 };
+
+pub async fn controller(
+    State(state): State<AppState>,
+    Json(req): Json<VerifyEmailService>,
+) -> Response {
+    let mut redis = match state.redis.get_multiplexed_tokio_connection().await {
+        Ok(redis) => redis,
+        Err(err) => {
+            return app_error_to_response(
+                AppError::infra(
+                    AppErrorKind::InternalError,
+                    "auth.controller.verify_email.redis",
+                    err,
+                )
+                .with_detail("Unable to connect to redis"),
+            );
+        },
+    };
+
+    match req.verify_email(&state.db, &mut redis).await {
+        Ok(_) => Response::new(ResponseCode::OK.into(), ResponseCode::OK.into(), None),
+        Err(err) => app_error_to_response(err),
+    }
+}
 
 #[derive(Deserialize)]
 pub struct VerifyEmailService {
