+++ auth/src/features/api/v1/common/mod.rs
@@ -1,0 +1,13 @@
+use axum::{Router, routing::any};
+
+use crate::{routers::cors, state::AppState};
+
+pub mod not_found;
+pub mod ping;
+
+pub fn router() -> Router<AppState> {
+    Router::new()
+        .route("/ping", any(ping::controller))
+        .fallback(not_found::controller)
+        .layer(cors::get_public_cors())
+}
