+++ auth/src/features/api/v1/common/not_found.rs
@@ -1,0 +1,5 @@
+use crate::routers::serializer::{Response, ResponseCode};
+
+pub async fn controller() -> Response {
+    ResponseCode::NotFound.into()
+}
