+++ auth/src/main.rs
@@ -3,10 +3,10 @@
 
 mod init;
 
+mod features;
 mod internal;
 mod models;
 mod routers;
-mod services;
 mod state;
 
 use std::{io, sync::Arc};
@@ -16,7 +16,7 @@ use colored::Colorize;
 use tokio::{net::TcpListener, signal, sync::oneshot};
 use tracing::{error, info};
 
-use crate::{internal::config::Config, routers::get_router};
+use crate::internal::config::Config;
 
 async fn shutdown_signal() {
     let ctrl_c = async {
@@ -87,8 +87,8 @@ async fn main() -> anyhow::Result<()> {
         })
         .await;
 
-    let app =
-        get_router(Router::new(), &config).with_state(state::APP_STATE.get().unwrap().clone());
+    let app = features::router(Router::new(), &config)
+        .with_state(state::APP_STATE.get().unwrap().clone());
 
     let addr = format!("{}:{}", &host, config.port);
