--- auth/src/routers/controllers.rs
@@ -1,4 +1,0 @@
-pub mod actions;
-pub mod auth;
-pub mod common;
-pub mod users;
