+++ auth/src/routers/mod.rs
@@ -1,0 +1,6 @@
+pub mod cors;
+pub mod extractor;
+pub mod middleware;
+pub mod response;
+pub mod serializer;
+pub mod utils;
