--- auth/src/services/actions/mod.rs
@@ -1,5 +1,0 @@
-pub mod reset_password;
-pub mod verify_email;
-
-pub use reset_password::*;
-pub use verify_email::*;
