--- auth/src/services/auth/mod.rs
@@ -1,11 +1,0 @@
-mod forget_password;
-mod login;
-mod register;
-mod reset_password;
-mod verify_email;
-
-pub use forget_password::*;
-pub use login::*;
-pub use register::*;
-pub use reset_password::*;
-pub use verify_email::*;
