--- auth/src/services/mod.rs
@@ -1,3 +1,0 @@
-pub mod actions;
-pub mod auth;
-pub mod users;
