--- auth/src/services/users/mod.rs
@@ -1,9 +1,0 @@
-mod delete;
-mod info;
-mod setting;
-mod update;
-
-pub use delete::*;
-pub use info::*;
-pub use setting::*;
-pub use update::*;
